Correcto funcionamiento de editar y administrar Recorridos y Reservas (Gracias a dios)

+++ src/seastar/views.py
@@ -4,7 +4,7 @@ from django.http import HttpRequest
 from django.contrib.auth.models import User
 from django.contrib import  messages
 from seastar.models import Navio, Itinerario, Puerto, Recorrido, Cubierta, Camarote, Pais, Ciudad, TipoDocumento, Pasajero, ReservaCamarote
-from datetime import date
+from datetime import datetime
 # Create your views here.
 
 def home(request:HttpRequest):
@@ -132,7 +132,7 @@ def profile(request: HttpRequest):
     if request.method == 'POST':
         messages.success(request,("Yippie"))
 
-        fecha = date.today()
+        fecha = datetime.today()
         camaroteReservado = Camarote.objects.get(numero_camarote=request.POST['seleccionCamarote'])
         nombreUsuario = user.username
         pasajero = Pasajero.objects.get(username=nombreUsuario)
@@ -140,7 +140,7 @@ def profile(request: HttpRequest):
         miReserva = ReservaCamarote.objects.create(fechaReserva=fecha,camaroteReservado=camaroteReservado,recorridoReservado=recorridoReservado)
         miReserva.listaPasajeros.add(pasajero)
         miReserva.save()
-        return redirect('./reserva.html')
+        return redirect('./profile.html')
     else :
         return render(request, "profile.html", {"logged_user": logged_user, "user" : user, "pasajero": pasajero})
 
@@ -152,23 +152,38 @@ def logout_view(request:HttpRequest):
 def adminRecorridos(request):
     logged_user = getLoggedUser(request)
     
-    
     recorridos = Recorrido.objects.all()
     navios = Navio.objects.all()
     itinerarios = Itinerario.objects.all()
     reservas = ReservaCamarote.objects.all()
     if request.method == 'POST':
-        num = request.POST['num']
-        iti = request.POST['itinerario']
-        nav = request.POST['navio']
-        fecha = request.POST['fecha']
-        duracion = request.POST['duracion']
+        if 'recorrido-borrar' in request.POST:
+            reser = request.POST['recorrido-borrar']
+            Recorrido.objects.get(id=reser).delete()
+        elif 'recorrido-guardar' in request.POST:
+            reser = request.POST['recorrido-guardar'] 
+            num = request.POST['num']
+            iti = request.POST['itinerario']
+            nav = request.POST['navio']
+            fecha = request.POST['fecha']
+            duracion = request.POST['duracion']
 
-        itiF = Itinerario.objects.get(nombreItinerario=iti)
-        navF = Navio.objects.get(nombreNavio=nav)
+            itiF = Itinerario.objects.get(nombreItinerario=iti)
+            navF = Navio.objects.get(nombreNavio=nav)
 
-        myRecorrido = Recorrido(numeroEscala=num,itinerarioRealizado=itiF,navioDelViaje=navF,fechaViaje=fecha,duracionViaje=duracion)
-        myRecorrido.save()
+            Recorrido.objects.filter(id=reser).update(numeroEscala=num,itinerarioRealizado=itiF,navioDelViaje=navF,fechaViaje=fecha,duracionViaje=duracion)
+        else: 
+            num = request.POST['num']
+            iti = request.POST['itinerario']
+            nav = request.POST['navio']
+            fecha = request.POST['fecha']
+            duracion = request.POST['duracion']
+
+            itiF = Itinerario.objects.get(nombreItinerario=iti)
+            navF = Navio.objects.get(nombreNavio=nav)
+
+            myRecorrido = Recorrido(numeroEscala=num,itinerarioRealizado=itiF,navioDelViaje=navF,fechaViaje=fecha,duracionViaje=duracion)
+            myRecorrido.save()
 
     return render(request, "adminRecorridos.html", { "itinerarios" : itinerarios , "reservas" : reservas , "navios" : navios , "recorridos": recorridos , "logged_user": logged_user })
 
@@ -178,16 +193,18 @@ def adminReserva(request):
     recorridos = Recorrido.objects.all()
     reservas = ReservaCamarote.objects.all()
     if request.method == "POST":
-        reser = request.POST['reserva']
-        ReservaCamarote.objects.get(id=reser).delete()
-    elif request.method == "PUT":
-        reserva = request.PUT['reserva']
-        camarote = request.PUT['camarote']
-        pasajeros = request.PUT['pasajeros']
-        print(reserva)
-        print(camarote)
-        print(pasajeros)
-        reserva.update(camaroteReservado=camarote,listaPasajeros=pasajeros)
+        if 'reserva-borrar' in request.POST:
+            reser = request.POST['reserva-borrar']
+            ReservaCamarote.objects.get(id=reser).delete()
+        elif 'reserva-guardar' in request.POST: 
+            reser = request.POST['reserva-guardar']
+            camarote = request.POST['camarote']
+            pasajeros = request.POST['pasajeros']
+            pasajerosF = Pasajero.objects.get(nombre=pasajeros)
+            camaroteF = Camarote.objects.get(numero_camarote=camarote)
+            ReservaCamarote.objects.filter(id=reser).update(camaroteReservado=camaroteF)
+            ReservaCamarote.objects.get(id=reser).listaPasajeros.clear()
+            ReservaCamarote.objects.get(id=reser).listaPasajeros.add(pasajerosF)
     return render(request, "adminReserva.html", { "recorridos": recorridos , "reservas" : reservas , "logged_user": logged_user })
 
 def editarRecorridos(request):
@@ -198,19 +215,10 @@ def editarRecorridos(request):
     itinerarios = Itinerario.objects.all()
     reservas = ReservaCamarote.objects.all()
     if request.method == 'POST':
-        num = request.POST['num']
-        iti = request.POST['itinerario']
-        nav = request.POST['navio']
-        fecha = request.POST['fecha']
-        duracion = request.POST['duracion']
-
-        itiF = Itinerario.objects.get(nombreItinerario=iti)
-        navF = Navio.objects.get(nombreNavio=nav)
-
-        myRecorrido = Recorrido(numeroEscala=num,itinerarioRealizado=itiF,navioDelViaje=navF,fechaViaje=fecha,duracionViaje=duracion)
-        myRecorrido.save()
-
-    return render(request, "editarRecorridos.html", { "itinerarios" : itinerarios , "reservas" : reservas , "navios" : navios , "recorridos": recorridos , "logged_user": logged_user })
+        recorrido = request.POST['recorrido']
+        recorridoElegido = Recorrido.objects.get(id=recorrido)
+        fechaYHora = recorridoElegido.fechaViaje.strftime("%Y-%m-%d %H:%M:%S")
+    return render(request, "editarRecorridos.html", { "itinerarios" : itinerarios , "reservas" : reservas , "navios" : navios , "recorridos": recorridos , "recorridoElegido" : recorridoElegido , "fechaYHora" : fechaYHora , "logged_user": logged_user })
 
 def editarReserva(request):
     logged_user = getLoggedUser(request)
@@ -221,7 +229,9 @@ def editarReserva(request):
     pasajeros = Pasajero.objects.all()
     if request.method == "POST":
         reser = request.POST['reserva']
-    return render(request, "editarReserva.html", { "recorridos": recorridos , "pasajeros": pasajeros , "reservas" : reservas , "camarotes" : camarotes , "logged_user": logged_user , "reservaElegida": reser })
+        reservaElegida = ReservaCamarote.objects.get(id=reser)
+        navio = reservaElegida.recorridoReservado.navioDelViaje.nombreNavio
+    return render(request, "editarReserva.html", { "recorridos": recorridos , "pasajeros": pasajeros , "reservas" : reservas , "camarotes" : camarotes , "logged_user" : logged_user , "reservaElegida" : reservaElegida , "navio" : navio })
 
 def getLoggedUser(request: HttpRequest):
     return request.session.get("user", "Iniciar Sesion")
